Saving to Mongo implemented in helper class
