--- SocketIOClient/index.html
@@ -1,118 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <title>WebSocket Video-Audio Streaming</title>
-</head>
-
-<body>
-    <p id="streamData"></p>
-    <video id="videoElement" autoplay></video>
-    <video id="videoElement2" autoplay></video>
-
-    <!-- Hier hardcode ik de room code die wordt gegenerate als iemand naar de server streamt, dit hoort userId te zijn uit de database (nu is het socket.id) -->
-    <button id="streamWatcher" onclick="joinRoom('1')">Join Room</button>
-    <button id="streamWatcher" onclick="joinRoom('roomName')">Join Room</button>
-    <script src="https://cdn.socket.io/4.0.0/socket.io.min.js"></script>
-    <script>
-        const socket = io("ws://localhost:3000/");
-        const video = document.getElementById("videoElement");
-        const video2 = document.getElementById("videoElement2");
-        let i = 0;
-
-        socket.on("connect", () => {
-            console.log("Connected to Socket.IO server");
-            startStreaming();
-        });
-
-        socket.on("disconnect", () => {
-            console.log("Disconnected from Socket.IO server");
-        });
-
-        socket.on("error", (error) => {
-            console.error("Socket.IO error:", error);
-        });
-
-
-        socket.on("stream", (streamObject) => {
-            console.log(streamObject);
-
-            // console.log(streamObject.sequenceId)
-            // const receiverBuffer = new Blob([streamObject.data], {
-            //     type: 'video/webm'
-            // });
-            // const videoUrl = URL.createObjectURL(receiverBuffer);
-            // video2.src = videoUrl;
-            // i++;
-            // const text = document.getElementById("streamData");
-            // text.textContent = `${streamObject.userId} ` + i;
-        });
-
-
-
-        function joinRoom(userId) {
-            socket.emit("joinRoom", userId);
-        }
-
-        function startStreaming() {
-            navigator.mediaDevices.getUserMedia({
-                video: true,
-                audio: true
-            })
-                .then(stream => {
-                    const mediaRecorder = new MediaRecorder(stream);
-
-                    mediaRecorder.ondataavailable = (event) => {
-                        if (event.data.size > 0 && socket.connected) {
-                            i++;
-                            // const header = generateVideoHeader(); // Generate header for each blob
-                            // const blobWithHeader = appendHeaderToBlob(header, event.data); // Append header to blob
-
-                            const streamObject = {
-                                data: event.data, // Send blob with header
-                                hash: "testHash", // Hash for the video stream
-                                userId: socket.id, // Change this to userId of streamers
-                                sequenceId: i, // Sequence ID for the video stream
-                            };
-                            // console.log(socket.id);
-                            socket.emit("stream", streamObject);
-                        }
-                    };
-
-                    video.srcObject = stream;
-                    // video2.srcObject = stream;
-                    mediaRecorder.start(1000); // Adjust the timeslice for data availability in ms
-                })
-                .catch(error => {
-                    console.error("Error accessing media devices.", error);
-                });
-
-
-        }
-
-        // Generate a header for the video
-        function generateVideoHeader() {
-            return {
-                type: 'video/webm', // Example MIME type for WebM video
-                codec: 'vp8', // Example video codec
-                resolution: {
-                    width: 1280,
-                    height: 720
-                }, // Example resolution
-                frameRate: 30, // Example frame rate
-                // Add any other metadata as needed
-            };
-        }
-
-        // Append header to the video blob
-        function appendHeaderToBlob(header, blob) {
-            const headerBlob = new Blob([JSON.stringify(header)], {
-                type: 'application/json'
-            });
-            return new Blob([headerBlob, blob]);
-        }
-    </script>
-
-</body>
-
-</html>
